Second bugfix pass
More indentation fixes.
Replaced "on fail" as they fired when the player cancelled the mission.
Changed "to fail" to "to complete"; otherwise, the player will only be
able to cancel under those conditions (not the intent).
Applied precis writing to all missions.
Removed branch from "Counter-Offer" as it didn't make sense at this
point in the story.
Paragarph indentations added. (See
https://github.com/endless-sky/endless-sky/issues/63 )